docs/vocabulary_development/scripts/requirements.txt
python-dotenv
networkx[default]
pydot